* bug fix PooledByteBufAllocator.expansion * add channelId(inc as JavaChannel) in EpollChannel * update max_write_backlog -> pending_write_size * update http codec * update ByteBuf pool.capacity >> memory size instead of unit count * rm Frame.setBytes * rm channel manager * rm ByteBuf.BufRecycle

+++ firenio-core/src/main/java/com/firenio/buffer/UnpooledByteBufAllocator.java
@@ -50,7 +50,7 @@ public final class UnpooledByteBufAllocator extends ByteBufAllocator {
     public void freeMemory() {}
 
     @Override
-    public int getCapacity() {
+    public long getCapacity() {
         return -1;
     }
